@@ -5,14 +5,16 @@ namespace Isap.Converters
 {
 	public abstract class ConvertAttempt
 	{
-		protected ConvertAttempt(bool isSuccess, object result)
+		protected ConvertAttempt(bool isSuccess, object result, string message)
 		{
 			IsSuccess = isSuccess;
 			Result = result;
+			Message = message;
 		}
 
 		public bool IsSuccess { get; }
 		public object Result { get; }
+		public string Message { get; }
 
 		public abstract Type ResultType { get; }
 
@@ -21,23 +23,23 @@ namespace Isap.Converters
 			return new ConvertAttempt<T>(true, result);
 		}
 
-		public static ConvertAttempt Success(object result)
+		public static ConvertAttempt Success(object result, string message = default)
 		{
 			if (result == null)
-				return new ConvertAttempt<object>(true, null);
+				return new ConvertAttempt<object>(true, null, message);
 			Type attemptType = typeof(ConvertAttempt<>).MakeGenericType(result.GetType());
-			return (ConvertAttempt) Activator.CreateInstance(attemptType, true, result);
+			return (ConvertAttempt) Activator.CreateInstance(attemptType, true, result, message);
 		}
 
-		public static ConvertAttempt<T> Fail<T>()
+		public static ConvertAttempt<T> Fail<T>(string message = default)
 		{
-			return new ConvertAttempt<T>(false, default(T));
+			return new ConvertAttempt<T>(false, default(T), message);
 		}
 
-		public static ConvertAttempt Fail(Type type)
+		public static ConvertAttempt Fail(Type type, string message = default)
 		{
 			Type attemptType = typeof(ConvertAttempt<>).MakeGenericType(type);
-			return (ConvertAttempt) Activator.CreateInstance(attemptType, false, attemptType.GetDefaultValue());
+			return (ConvertAttempt) Activator.CreateInstance(attemptType, false, attemptType.GetDefaultValue(), message);
 		}
 
 		public ConvertAttempt<T> Cast<T>()
@@ -48,8 +50,8 @@ namespace Isap.Converters
 
 	public class ConvertAttempt<T>: ConvertAttempt
 	{
-		public ConvertAttempt(bool isSuccess, T result)
-			: base(isSuccess, result)
+		public ConvertAttempt(bool isSuccess, T result, string message = default)
+			: base(isSuccess, result, message)
 		{
 		}
 
