@@ -1,26 +1,40 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
+using IdentityModel;
 using Isap.Abp.BackgroundJobs.EntityFrameworkCore.PostgreSql;
+using Isap.Abp.Extensions;
 using Isap.Abp.Extensions.Logging;
+using Isap.Abp.Extensions.MultiTenancy;
+using Isap.Abp.Extensions.Web;
 using Isap.CommonCore.Integrations;
 using Isap.CommonCore.Web.Middlewares.RequestLogging;
 using Isap.CommonCore.Web.Middlewares.Tracing;
 using Isap.Converters;
+using Microsoft.AspNetCore.Authentication.OAuth.Claims;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.IdentityModel.Protocols.OpenIdConnect;
 using MyCompanyName.MyProjectName.EntityFrameworkCore;
 using MyCompanyName.MyProjectName.Localization;
 using MyCompanyName.MyProjectName.MultiTenancy;
 using MyCompanyName.MyProjectName.Web.Menus;
 using Microsoft.OpenApi.Models;
+using StackExchange.Redis;
 using Volo.Abp;
 using Volo.Abp.Account.Web;
 using Volo.Abp.AspNetCore.Authentication.JwtBearer;
+using Volo.Abp.AspNetCore.Authentication.OpenIdConnect;
+using Volo.Abp.AspNetCore.MultiTenancy;
 using Volo.Abp.AspNetCore.Mvc;
+using Volo.Abp.AspNetCore.Mvc.Client;
 using Volo.Abp.AspNetCore.Mvc.Localization;
 using Volo.Abp.AspNetCore.Mvc.UI;
 using Volo.Abp.AspNetCore.Mvc.UI.Bootstrap;
@@ -29,14 +43,21 @@ using Volo.Abp.AspNetCore.Mvc.UI.MultiTenancy;
 using Volo.Abp.AspNetCore.Mvc.UI.Theme.Basic;
 using Volo.Abp.AspNetCore.Mvc.UI.Theme.Basic.Bundling;
 using Volo.Abp.AspNetCore.Mvc.UI.Theme.Shared;
+using Volo.Abp.AspNetCore.Security.Claims;
 using Volo.Abp.AspNetCore.Serilog;
 using Volo.Abp.Autofac;
 using Volo.Abp.AutoMapper;
+using Volo.Abp.Caching;
+using Volo.Abp.Caching.StackExchangeRedis;
 using Volo.Abp.FeatureManagement;
+using Volo.Abp.Http.Client.IdentityModel.Web;
 using Volo.Abp.Identity.Web;
 using Volo.Abp.Localization;
 using Volo.Abp.Modularity;
+using Volo.Abp.MultiTenancy;
+using Volo.Abp.MultiTenancy.ConfigurationStore;
 using Volo.Abp.PermissionManagement.Web;
+using Volo.Abp.Security.Claims;
 using Volo.Abp.Swashbuckle;
 using Volo.Abp.TenantManagement.Web;
 using Volo.Abp.UI.Navigation.Urls;
@@ -53,8 +74,15 @@ namespace MyCompanyName.MyProjectName.Web
         typeof(MyProjectNameHttpApiModule),
         typeof(MyProjectNameApplicationModule),
         typeof(MyProjectNameEntityFrameworkCoreDbMigrationsModule),
+        typeof(IsapAbpExtensionsRemoteSettingsModule),
+        typeof(IsapAbpExtensionsWebModule),
         typeof(IsapAbpBackgroundJobsPostgreSqlModule),
+        typeof(AbpAspNetCoreAuthenticationOpenIdConnectModule),
+        typeof(AbpAspNetCoreMvcClientModule),
         typeof(AbpAutofacModule),
+        typeof(AbpCachingStackExchangeRedisModule),
+        typeof(AbpFeatureManagementWebModule),
+        typeof(AbpHttpClientIdentityModelWebModule),
         typeof(AbpIdentityWebModule),
         typeof(AbpAccountWebIdentityServerModule),
         typeof(AbpAspNetCoreMvcUiBasicThemeModule),
@@ -65,6 +93,8 @@ namespace MyCompanyName.MyProjectName.Web
         )]
     public class MyProjectNameWebModule : AbpModule
     {
+        private const string DefaultCorsPolicyName = "Default";
+
         public override void PreConfigureServices(ServiceConfigurationContext context)
         {
             context.Services.PreConfigure<AbpMvcDataAnnotationsLocalizationOptions>(options =>
@@ -89,14 +119,20 @@ namespace MyCompanyName.MyProjectName.Web
             context.Services.AddSingleton(converter);
 
             ConfigureRequestLogging(context, converter, configuration);
+            ConfigureForwardedHeaders(context);
 
+            ConfigureSession(context, configuration);
             ConfigureUrls(configuration);
             ConfigureBundles();
+            ConfigureCache(configuration);
+            ConfigureRedis(context, configuration, hostingEnvironment);
             ConfigureAuthentication(context, configuration);
             ConfigureAutoMapper();
             ConfigureVirtualFileSystem(hostingEnvironment);
             ConfigureLocalizationServices();
             ConfigureNavigationServices();
+            ConfigureMultiTenancy(context, converter, configuration);
+            ConfigureCors(context, configuration);
             ConfigureAutoApiControllers();
             ConfigureSwaggerServices(context.Services);
         }
@@ -113,6 +149,31 @@ namespace MyCompanyName.MyProjectName.Web
 
                     List<string> basePaths = config.GetValue("BasePaths", () => new List<string>());
                     options.AddBasePaths(basePaths.ToArray());
+                });
+        }
+
+        private void ConfigureForwardedHeaders(ServiceConfigurationContext context)
+        {
+            context.Services.Configure<ForwardedHeadersOptions>(options =>
+                {
+                    options.ForwardedHeaders = ForwardedHeaders.All;
+                    options.KnownProxies.Clear();
+                    options.KnownNetworks.Clear();
+                });
+        }
+
+        private void ConfigureSession(ServiceConfigurationContext context, IConfiguration configuration)
+        {
+            // Add Distributed Redis Cache for Session
+            context.Services.AddStackExchangeRedisCache(options =>
+                {
+                    options.Configuration = configuration["Redis:Configuration"] ?? "localhost";
+                    options.InstanceName = "Session_";
+                });
+            context.Services.AddSession(options =>
+                {
+                    // 20 minutes later from last access your session will be removed.
+                    options.IdleTimeout = TimeSpan.FromMinutes(20);
                 });
         }
 
@@ -138,14 +199,90 @@ namespace MyCompanyName.MyProjectName.Web
             });
         }
 
+        private void ConfigureCache(IConfiguration configuration)
+        {
+            Configure<AbpDistributedCacheOptions>(options =>
+                {
+                    options.KeyPrefix = "MyProjectName:";
+                });
+        }
+
+        private void ConfigureMultiTenancy(ServiceConfigurationContext context, IValueConverter converter, IConfiguration configuration)
+        {
+            Configure<AbpMultiTenancyOptions>(options =>
+                {
+                    options.IsEnabled = MultiTenancyConsts.IsEnabled;
+                });
+
+            Configure<AbpExtWebOptions>(configuration.GetSection("AbpExtWebOptions"));
+
+            if (MultiTenancyConsts.IsEnabled)
+            {
+                Configure<AbpTenantResolveOptions>(options =>
+                    {
+                        options.TenantResolvers.Add(new DomainNameTenantResolveContributor());
+                    });
+
+                Configure<AbpDefaultTenantStoreOptions>(options =>
+                    {
+                        options.Tenants = MultiTenancyConsts.DefaultTenants;
+                    });
+
+                Configure<AbpAspNetCoreMultiTenancyOptions>(options =>
+                    {
+                        options.TenantKey = IsapMultiTenancyConsts.TenantHeaderName;
+                    });
+            }
+        }
+
         private void ConfigureAuthentication(ServiceConfigurationContext context, IConfiguration configuration)
         {
             context.Services.AddAuthentication()
                 .AddJwtBearer(options =>
+                    {
+                        options.Authority = configuration["AuthServer:Authority"];
+                        options.RequireHttpsMetadata = Convert.ToBoolean(configuration["AuthServer:RequireHttpsMetadata"]);
+                        options.Audience = "MyProjectName";
+                    });
+            context.Services
+                .AddAuthentication(options =>
+                    {
+                        options.DefaultScheme = "Cookies";
+                        options.DefaultChallengeScheme = "oidc";
+                    })
+                .AddCookie("Cookies", options =>
+                    {
+                        options.ExpireTimeSpan = TimeSpan.FromDays(365);
+                    })
+                .AddAbpOpenIdConnect("oidc", options =>
+                    {
+                        options.Authority = configuration["AuthServer:Authority"];
+                        options.RequireHttpsMetadata = Convert.ToBoolean(configuration["AuthServer:RequireHttpsMetadata"]);
+                        options.ResponseType = OpenIdConnectResponseType.CodeIdToken;
+
+                        options.ClientId = configuration["AuthServer:ClientId"];
+                        options.ClientSecret = configuration["AuthServer:ClientSecret"];
+
+                        options.SaveTokens = true;
+                        options.GetClaimsFromUserInfoEndpoint = true;
+
+                        options.Scope.Add("role");
+                        options.Scope.Add("email");
+                        options.Scope.Add("phone");
+                        options.Scope.Add("Vicon");
+
+                        options.ClaimActions.MapAbpClaimTypes();
+                    })
+                ;
+
+            Configure<AbpClaimsMapOptions>(options =>
                 {
-                    options.Authority = configuration["AuthServer:Authority"];
-                    options.RequireHttpsMetadata = Convert.ToBoolean(configuration["AuthServer:RequireHttpsMetadata"]);
-                    options.Audience = "MyProjectName";
+                    //options.Maps.Add(JwtClaimTypes.Subject, () => AbpClaimTypes.UserId);
+                    //options.Maps.Add(JwtClaimTypes.PreferredUserName, () => AbpClaimTypes.UserName);
+                    //options.Maps.Add(JwtClaimTypes.Email, () => AbpClaimTypes.Email);
+                    options.Maps.Add(JwtClaimTypes.EmailVerified, () => AbpClaimTypes.EmailVerified);
+                    options.Maps.Add(JwtClaimTypes.PhoneNumber, () => AbpClaimTypes.PhoneNumber);
+                    options.Maps.Add(JwtClaimTypes.PhoneNumberVerified, () => AbpClaimTypes.PhoneNumberVerified);
                 });
         }
 
@@ -232,6 +369,42 @@ namespace MyCompanyName.MyProjectName.Web
             );
         }
 
+        private void ConfigureRedis(
+            ServiceConfigurationContext context,
+            IConfiguration configuration,
+            IWebHostEnvironment hostingEnvironment)
+        {
+            if (!hostingEnvironment.IsDevelopment())
+            {
+                var redis = ConnectionMultiplexer.Connect(configuration["Redis:Configuration"]);
+                context.Services
+                    .AddDataProtection()
+                    .PersistKeysToStackExchangeRedis(redis, "MyProjectName-Protection-Keys");
+            }
+        }
+
+        private void ConfigureCors(ServiceConfigurationContext context, IConfiguration configuration)
+        {
+            context.Services.AddCors(options =>
+                {
+                    options.AddPolicy(DefaultCorsPolicyName, builder =>
+                        {
+                            builder
+                                .WithOrigins(
+                                    configuration["App:CorsOrigins"]
+                                        .Split(",", StringSplitOptions.RemoveEmptyEntries)
+                                        .Select(o => o.RemovePostFix("/"))
+                                        .ToArray()
+                                )
+                                .WithAbpExposedHeaders()
+                                .SetIsOriginAllowedToAllowWildcardSubdomains()
+                                .AllowAnyHeader()
+                                .AllowAnyMethod()
+                                .AllowCredentials();
+                        });
+                });
+        }
+
         public override void OnApplicationInitialization(ApplicationInitializationContext context)
         {
             var app = context.GetApplicationBuilder();
@@ -240,9 +413,12 @@ namespace MyCompanyName.MyProjectName.Web
             app.UseMiddleware<LoggingTraceIdentifierMiddleware>();
             app.UseRequestResponseLogging();
 
+            app.UseHttpMethodOverride();
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseForwardedHeaders();
             }
 
             app.UseAbpRequestLocalization();
@@ -250,13 +426,17 @@ namespace MyCompanyName.MyProjectName.Web
             if (!env.IsDevelopment())
             {
                 app.UseErrorPage();
+                app.UseForwardedHeaders();
             }
 
             app.UseCorrelationId();
             app.UseVirtualFiles();
+            app.UseSession();
             app.UseRouting();
+            app.UseCors(DefaultCorsPolicyName);
             app.UseAuthentication();
             app.UseJwtTokenMiddleware();
+            app.UseAbpClaimsMap();
 
             if (MultiTenancyConsts.IsEnabled)
             {
@@ -267,9 +447,9 @@ namespace MyCompanyName.MyProjectName.Web
             app.UseAuthorization();
             app.UseSwagger();
             app.UseAbpSwaggerUI(options =>
-            {
-                options.SwaggerEndpoint("/swagger/v1/swagger.json", "MyProjectName API");
-            });
+                {
+                    options.SwaggerEndpoint("/swagger/v1/swagger.json", "MyProjectName API");
+                });
             app.UseAuditing();
             app.UseAbpSerilogEnrichers();
             app.UseConfiguredEndpoints();
